Release a new version of the debugger (#3826)
