+++ configs/firefox-panel.json
@@ -34,6 +34,9 @@
     },
     "outline": {
       "enabled": true
+    },
+    "wasm": {
+      "enabled": true
     }
   }
 }
