+++ src/utils/parser/utils/helpers.js
@@ -3,8 +3,6 @@
 import * as t from "babel-types";
 import type { NodePath, Node } from "babel-traverse";
 
-import type { AstLocation, AstPosition } from "../types";
-
 export function isLexicalScope(path: NodePath) {
   return t.isBlockStatement(path) || isFunction(path) || t.isProgram(path);
 }
@@ -54,22 +52,4 @@ export function getMemberExpression(root: Node) {
 
   const expr = _getMemberExpression(root, []);
   return expr.join(".");
-}
-
-export function containsPosition(a: AstLocation, b: AstPosition) {
-  const startsBefore =
-    a.start.line < b.line ||
-    (a.start.line === b.line && a.start.column <= b.column);
-  const endsAfter =
-    a.end.line > b.line || (a.end.line === b.line && a.end.column >= b.column);
-
-  return startsBefore && endsAfter;
-}
-
-export function containsLocation(a: AstLocation, b: AstLocation) {
-  return containsPosition(a, b.start) && containsPosition(a, b.end);
-}
-
-export function nodeContainsPosition(node: Node, position: AstPosition) {
-  return containsPosition(node.loc, position);
 }
